contenereized/streamlit_app/requirements.txt
streamlit==1.32.2
streamlit-option-menu==0.3.12
requests==2.31.0
mysql-connector-python==8.3.0
pandas==2.2.1
pillow==10.2.0